requirements.txt
streamlit
requests
textblob
python-dotenv
google-genai